instructional_ai_system/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-jose[cryptography]
passlib[bcrypt]
python-multipart
groq
python-dotenv
PyPDF2
python-docx
openpyxl
youtube-transcript-api
beautifulsoup4
requests
python-dateutil
tenacity
json-repair
fastapi-mail
